fixed clippy, added windows subsystem

+++ src/options.rs
@@ -35,7 +35,7 @@ pub static GAME_NAME: &str = "Flappy Bevy";
         pub static PLAYER_START_Y: f32 = -200.0;
 
     // clouds
-        pub static CLOUDS_START_X: f32 = -SCREEN_X_BOUNDARY - SPRITE_SIZE * 3 as f32;
+        pub static CLOUDS_START_X: f32 = -SCREEN_X_BOUNDARY - SPRITE_SIZE * 3_f32;
         pub static CLOUDS_GAP_BETWEEN: f32 = 250.0;
         pub static CLOUDS_SPEED: f32 = 0.7;
         pub static CLOUDS_Y_RANGE: [f32; 2] = [-150.0, 300.0];
